Implement the search function
This implements search with and without Javascript enabled.
The search is always done on the server. Search results get fetched asynchronously on the client when JS is available and synchronously when no JS is available.

It also moves generating the slugs for recipes onto the server side completely, which decreases the bundle size.

Furthermore, a proper icon is created.

This closes #5
This closes #6